+++ app/lib/types.ts
@@ -1,0 +1,9 @@
+import type { LambdaErrorInfo } from '@remotion/lambda';
+
+export type RenderStatusResponse = {
+	renderId: string;
+	done: boolean;
+	overallProgress: number;
+	outputFile: string | null;
+	errors: LambdaErrorInfo[];
+}[];
